requirements.txt
flask
Gunicorn
google-cloud-dialogflow
requests
pusher
SpeechRecognition
pyaudio
python-dotenv